aws/pizzaluvrs/requirements.txt
Django==2.2.5
django-extensions==2.2.5
django-filter==2.2.0
djangorestframework==3.10.3
Markdown==3.1.1
Pillow==6.1.0
psycopg2-binary==2.8.3
python-dotenv==0.10.3
pytz==2019.2
six==1.13.0
sqlparse==0.3.0
